chore: separate handlers from boilerplate
